+++ .examples/backend/src/endpoint/views/home/home.module.ts
@@ -1,4 +1,4 @@
-import { CacheModule, Module, ConfigModule } from "@zum-front-core/backend";
+import { CacheModule, Module, ConfigModule } from "@zum-front-end/backend";
 import { HomeController } from "./home.controller";
 import { HomeService } from "./home.service";